Extract TaskKeywordResolver from EventResolverBase
